peekingduck/training/src/requirements.txt
ipykernel
nbformat
black
pandas
numpy
hydra-core
scikit-learn
pytorch
torchvision
tensorflow
